Move the placeholder leaf down, and give the two prefixes distinct names
Plan 320 E2.1. A guest never holds a credential -- it holds an opaque
per-session token and routes the request to a host-local substitution
endpoint. What that token looks like, and how to find one in a header,
is pure: no custody, no I/O, no randomness. It moves to mvm-contract so
the browser demo runs the host's scan rather than a second copy.

Moved: the reserved prefix, the `Placeholder` newtype with `as_str`, and
`find_placeholder`, plus the two tests that cover the scan.

Stayed: minting. Generating a token draws from the OS RNG, which is the
`getrandom`-in-the-bundle dependency the browser build exists to avoid.
`Placeholder::new` takes the token as given, so the caller owns where the
bytes came from -- and its doc says plainly that the type is a wrapper,
not a guarantee of entropy.

The rename is the part worth reviewing. mvm-contract already had a
`PLACEHOLDER_PREFIX` -- `"mvm-managed:"`, the CLI binding's display form
in `policy::secret_binding`. Moving a second constant in under the same
name would have left two unrelated reserved prefixes in one crate,
distinguishable only by module path: the exact shape of the `AuditEntry`
collision that E3 is now blocked on untangling. So the moved one lands as
`SECRET_PLACEHOLDER_PREFIX` and its four call sites are updated. Hard
rename, no alias, matching how Increment 3 resolved `NetworkPolicy`.

A test asserts the two constants differ and that `find_placeholder`
returns `None` on an `mvm-managed:` string, so the distinction is checked
rather than merely documented. The module header tabulates all three
notations -- including `${NAME}`, which is authoring-only and which
nothing resolves at runtime.

Side effect: the placeholder scan now runs under wasm32-wasip1 for the
first time (653 -> 656).
